refactor: move calc functions into new tooltip info file
